pyproject.toml
[tool.poetry]
name = "aeris-hil"
version = "0.1.0"
description = "Advanced Enterprise-Ready Integration & Simulation – HIL Automation & Traceability Framework"
authors = ["Daniel Mondragón"]
readme = "README.md"
packages = [{ include = "aeris_hil" }]

[tool.poetry.dependencies]
python = "^3.11"
robotframework = "^6.1"
python-can = "^4.3"
pandas = "^2.2"
jinja2 = "^3.1"
flask = "^3.0"
requests = "^2.31"
jira = "^3.10.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
black = "^24.0"
ruff = "^0.3.0"
coverage = "^7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

